[expo-router] refactor link preview to show components in docs
