+++ packages/expo-router/src/exports.ts
@@ -15,7 +15,8 @@ export {
 
 export { router, Router } from './imperative-api';
 
-export { Link, Redirect, RedirectProps, LinkProps, WebAnchorProps } from './link/Link';
+export * from './link/Link';
+export * from './link/elements';
 export { useIsPreview } from './link/preview/PreviewRouteContext';
 
 export { withLayoutContext } from './layouts/withLayoutContext';
