+++ packages/expo-router/src/link/InternalLinkPreviewContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+import type { LinkProps } from './useLinkHooks';
+
+export const InternalLinkPreviewContext = createContext<
+  { isVisible: boolean; href: LinkProps['href'] } | undefined
+>(undefined);
