+++ packages/expo-router/src/link/LinkWithPreview.tsx
@@ -1,38 +1,23 @@
 'use client';
 
 import React, {
-  createContext,
   isValidElement,
-  use,
   useEffect,
   useMemo,
   useRef,
   useState,
-  type FC,
-  type PropsWithChildren,
   type ReactElement,
 } from 'react';
-import type { SFSymbol } from 'sf-symbols-typescript';
 
 import { useRouter } from '../hooks';
 import { BaseExpoRouterLink } from './BaseExpoRouterLink';
-import { HrefPreview } from './preview/HrefPreview';
+import { InternalLinkPreviewContext } from './InternalLinkPreviewContext';
+import { LinkMenu, LinkPreview, LinkTrigger } from './elements';
 import { useLinkPreviewContext } from './preview/LinkPreviewContext';
-import {
-  NativeLinkPreview,
-  NativeLinkPreviewAction,
-  NativeLinkPreviewContent,
-  NativeLinkPreviewTrigger,
-} from './preview/native';
+import { NativeLinkPreview, NativeLinkPreviewTrigger } from './preview/native';
 import { useNextScreenId } from './preview/useNextScreenId';
 import { LinkProps } from './useLinkHooks';
 import { shouldLinkExternally } from '../utils/url';
-import { useIsPreview } from './preview/PreviewRouteContext';
-import { Slot } from '../ui/Slot';
-
-const InternalLinkPreviewContext = createContext<
-  { isVisible: boolean; href: LinkProps['href'] } | undefined
->(undefined);
 
 export function LinkWithPreview({ children, ...rest }: LinkProps) {
   const router = useRouter();
@@ -145,168 +130,6 @@ export function LinkWithPreview({ children, ...rest }: LinkProps) {
       </InternalLinkPreviewContext>
     </NativeLinkPreview>
   );
-}
-
-interface LinkMenuActionProps {
-  /**
-   * The title of the menu item.
-   */
-  title: string;
-  /**
-   * Optional SF Symbol displayed alongside the menu item.
-   */
-  icon?: SFSymbol;
-  /**
-   * If true, the menu item will be disabled and not selectable.
-   *
-   * https://developer.apple.com/documentation/uikit/uimenuelement/attributes/disabled
-   */
-  disabled?: boolean;
-  /**
-   * If true, the menu item will be displayed as destructive.
-   *
-   * https://developer.apple.com/documentation/uikit/uimenuelement/attributes/destructive
-   */
-  destructive?: boolean;
-  /**
-   * If true, the menu will be kept presented after the action is selected.
-   *
-   * This is marked as unstable, because when action is selected it will recreate the menu,
-   * which will close all opened submenus and reset the scroll position.
-   *
-   * https://developer.apple.com/documentation/uikit/uimenuelement/attributes/keepsmenupresented
-   */
-  unstable_keepPresented?: boolean;
-  /**
-   * If true, the menu item will be displayed as selected.
-   */
-  isOn?: boolean;
-  onPress: () => void;
-}
-
-export function LinkMenuAction(props: LinkMenuActionProps) {
-  if (useIsPreview() || process.env.EXPO_OS !== 'ios' || !use(InternalLinkPreviewContext)) {
-    return null;
-  }
-  const { unstable_keepPresented, onPress, ...rest } = props;
-  return (
-    <NativeLinkPreviewAction
-      {...rest}
-      onSelected={onPress}
-      keepPresented={unstable_keepPresented}
-    />
-  );
-}
-
-export interface LinkMenuProps {
-  /**
-   * The title of the menu item
-   */
-  title?: string;
-  /**
-   * Optional SF Symbol displayed alongside the menu item.
-   */
-  icon?: string;
-  /**
-   * If true, the menu will be displayed as a palette.
-   * This means that the menu will be displayed as one row
-   *
-   * https://developer.apple.com/documentation/uikit/uimenu/options-swift.struct/displayaspalette
-   */
-  displayAsPalette?: boolean;
-  /**
-   * If true, the menu will be displayed inline.
-   * This means that the menu will not be collapsed
-   *
-   * https://developer.apple.com/documentation/uikit/uimenu/options-swift.struct/displayinline
-   */
-  displayInline?: boolean;
-  /**
-   * If true, the menu item will be displayed as destructive.
-   *
-   * https://developer.apple.com/documentation/uikit/uimenu/options-swift.struct/destructive
-   */
-  destructive?: boolean;
-  children: ReactElement<LinkMenuActionProps> | ReactElement<LinkMenuActionProps>[];
-}
-
-export const LinkMenu: FC<LinkMenuProps> = (props) => {
-  if (useIsPreview() || process.env.EXPO_OS !== 'ios' || !use(InternalLinkPreviewContext)) {
-    return null;
-  }
-  const children = React.Children.toArray(props.children).filter(
-    (child) => isValidElement(child) && (child.type === LinkMenuAction || child.type === LinkMenu)
-  );
-  return (
-    <NativeLinkPreviewAction
-      {...props}
-      title={props.title ?? ''}
-      onSelected={() => {}}
-      children={children}
-    />
-  );
-};
-
-interface LinkPreviewProps {
-  /**
-   * Sets the preferred width of the preview.
-   * If not set, full width of the screen will be used.
-   *
-   * This is only **preferred** width, the actual width may be different
-   */
-  width?: number;
-
-  /**
-   * Sets the preferred height of the preview.
-   * If not set, full height of the screen will be used.
-   *
-   * This is only **preferred** height, the actual height may be different
-   */
-  height?: number;
-  children?: React.ReactNode;
-}
-
-export function LinkPreview({ children, width, height }: LinkPreviewProps) {
-  const internalPreviewContext = use(InternalLinkPreviewContext);
-  if (useIsPreview() || process.env.EXPO_OS !== 'ios' || !internalPreviewContext) {
-    return null;
-  }
-  const { isVisible, href } = internalPreviewContext;
-  const contentSize = {
-    width: width ?? 0,
-    height: height ?? 0,
-  };
-  let content: React.ReactNode;
-  if (children) {
-    content = isVisible ? children : null;
-  } else {
-    content = isVisible ? <HrefPreview href={href} /> : null;
-  }
-
-  return (
-    <NativeLinkPreviewContent
-      style={{
-        /* Setting default background here, so that the preview is not transparent */
-        backgroundColor: '#fff',
-      }}
-      preferredContentSize={contentSize}>
-      {content}
-    </NativeLinkPreviewContent>
-  );
-}
-
-export function LinkTrigger(props: PropsWithChildren) {
-  if (React.Children.count(props.children) > 1 || !isValidElement(props.children)) {
-    // If onPress is passed, this means that Link passed props to this component.
-    // We can assume that asChild is used, so we throw an error, because link will not work in this case.
-    if (props && typeof props === 'object' && 'onPress' in props) {
-      throw new Error(
-        'When using Link.Trigger in an asChild Link, you must pass a single child element that will emit onPress event.'
-      );
-    }
-    return props.children;
-  }
-  return <Slot {...props} />;
 }
 
 function getFirstChildOfType<PropsT>(
